move scrollbar into carousel
